requirements.txt
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.2
xgboost==2.0.0
matplotlib==3.8.0
seaborn==0.13.0
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
pydantic==2.8.0
joblib==1.3.2
imbalanced-learn==0.11.0
